Split individual transaction specs into separate files
